# Metrics: Improve perf by moving timestamps off of MetricPoint