--- a/src/OpenTelemetry/Metrics/MetricPointsAccessor.cs
+++ b/src/OpenTelemetry/Metrics/MetricPointsAccessor.cs
@@ -14,7 +14,6 @@
 // limitations under the License.
 // </copyright>
 
-using System;
 using OpenTelemetry.Internal;
 
 namespace OpenTelemetry.Metrics
@@ -28,18 +27,14 @@ namespace OpenTelemetry.Metrics
         private readonly MetricPoint[] metricsPoints;
         private readonly int[] metricPointsToProcess;
         private readonly long targetCount;
-        private readonly DateTimeOffset start;
-        private readonly DateTimeOffset end;
 
-        internal MetricPointsAccessor(MetricPoint[] metricsPoints, int[] metricPointsToProcess, long targetCount, DateTimeOffset start, DateTimeOffset end)
+        internal MetricPointsAccessor(MetricPoint[] metricsPoints, int[] metricPointsToProcess, long targetCount)
         {
             Guard.ThrowIfNull(metricsPoints);
 
             this.metricsPoints = metricsPoints;
             this.metricPointsToProcess = metricPointsToProcess;
             this.targetCount = targetCount;
-            this.start = start;
-            this.end = end;
         }
 
         /// <summary>
@@ -47,9 +42,7 @@ namespace OpenTelemetry.Metrics
         /// </summary>
         /// <returns><see cref="Enumerator"/>.</returns>
         public Enumerator GetEnumerator()
-        {
-            return new Enumerator(this.metricsPoints, this.metricPointsToProcess, this.targetCount, this.start, this.end);
-        }
+            => new(this.metricsPoints, this.metricPointsToProcess, this.targetCount);
 
         /// <summary>
         /// Enumerates the elements of a <see cref="MetricPointsAccessor"/>.
@@ -58,31 +51,22 @@ namespace OpenTelemetry.Metrics
         {
             private readonly MetricPoint[] metricsPoints;
             private readonly int[] metricPointsToProcess;
-            private readonly DateTimeOffset start;
-            private readonly DateTimeOffset end;
             private readonly long targetCount;
             private long index;
 
-            internal Enumerator(MetricPoint[] metricsPoints, int[] metricPointsToProcess, long targetCount, DateTimeOffset start, DateTimeOffset end)
+            internal Enumerator(MetricPoint[] metricsPoints, int[] metricPointsToProcess, long targetCount)
             {
                 this.metricsPoints = metricsPoints;
                 this.metricPointsToProcess = metricPointsToProcess;
                 this.targetCount = targetCount;
                 this.index = -1;
-                this.start = start;
-                this.end = end;
             }
 
             /// <summary>
             /// Gets the <see cref="MetricPoint"/> at the current position of the enumerator.
             /// </summary>
             public ref readonly MetricPoint Current
-            {
-                get
-                {
-                    return ref this.metricsPoints[this.metricPointsToProcess[this.index]];
-                }
-            }
+                => ref this.metricsPoints[this.metricPointsToProcess[this.index]];
 
             /// <summary>
             /// Advances the enumerator to the next element of the <see
@@ -93,17 +77,7 @@ namespace OpenTelemetry.Metrics
             /// langword="false"/> if the enumerator has passed the end of the
             /// collection.</returns>
             public bool MoveNext()
-            {
-                while (++this.index < this.targetCount)
-                {
-                    ref var metricPoint = ref this.metricsPoints[this.metricPointsToProcess[this.index]];
-                    metricPoint.StartTime = this.start;
-                    metricPoint.EndTime = this.end;
-                    return true;
-                }
-
-                return false;
-            }
+                => ++this.index < this.targetCount;
         }
     }
 }