## Changes

Instead of mirroring Start + End times on `MetricPoint`, store a reference back to `AggregatorStore`. This accomplishes two things:

* Reduces the size of `MetricPoint` from 104B to 80B. This lowers the overall memory footprint of the metrics pipeline.
* Speeds up export enumeration which previously was managing these values in `MoveNext`. Enumerator struct is also smaller which will slightly speed up creating the enumerator.

## Benchmarks

### Before

|  Method | UseWithRef |     Mean |    Error |   StdDev |   Median |  Gen 0 | Allocated |
|-------- |----------- |---------:|---------:|---------:|---------:|-------:|----------:|
| Collect |       True | 26.19 us | 0.534 us | 1.567 us | 26.69 us | 0.0305 |     160 B |

### After

|  Method | UseWithRef |     Mean |    Error |   StdDev |  Gen 0 | Allocated |
|-------- |----------- |---------:|---------:|---------:|-------:|----------:|
| Collect |       True | **18.01 us** | 0.529 us | 1.551 us | 0.0305 |     160 B |

## TODOs

* [ ] Appropriate `CHANGELOG.md` updated for non-trivial changes