@CodeBlanch 2022-03-29
@alanwest 

> That said, if someone kept a handle on a Metric object somehow, the results of calling GetMetricPoints is meaningless outside of the expected flow.

Totally!

We have been discussing that a bit here: https://github.com/open-telemetry/opentelemetry-dotnet/pull/3071#discussion_r832686083

Current thinking is, for general metrics export, we will publish that it is explicitly not safe to capture `Metric` for use after the `Export` call. In the case of in-memory exporter, we can introduce a class to store the data and make it safe to store.